Models naming & type changes

+++ SharedModels/Customer.cs
@@ -6,7 +6,7 @@ namespace SharedModels {
     public class Customer {
         [Key]
         public Guid Id { get; set; } = Guid.NewGuid();
-        public int Balance { get; set; }
+        public long Balance { get; set; }
 
         public Guid UserId { get; set; }
         [ForeignKey("UserId")]
